feat: add support for CURL/PULLUP exercises and fixed backend environment/schema blockers

+++ database/schema.sql
@@ -1,6 +1,6 @@
 -- User details and authentication
 CREATE TABLE users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     email VARCHAR(255) UNIQUE NOT NULL,
     password VARCHAR(255) NOT NULL,
@@ -10,8 +10,8 @@ CREATE TABLE users (
 
 -- Workout sessions tracker
 CREATE TABLE workout_sessions (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT REFERENCES users(id) ON DELETE CASCADE,
     exercise_type VARCHAR(100) NOT NULL, -- e.g. PUSHUP, SQUAT
     start_time TIMESTAMP NOT NULL,
     end_time TIMESTAMP,
@@ -20,8 +20,8 @@ CREATE TABLE workout_sessions (
 
 -- Per-rep or aggregated exercise metrics
 CREATE TABLE exercise_metrics (
-    id SERIAL PRIMARY KEY,
-    session_id INTEGER REFERENCES workout_sessions(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    session_id BIGINT REFERENCES workout_sessions(id) ON DELETE CASCADE,
     reps INTEGER DEFAULT 0,
     avg_speed NUMERIC(5,2), -- in seconds per rep
     accuracy NUMERIC(5,2), -- percentage of 'GOOD' posture out of total reps
@@ -30,8 +30,8 @@ CREATE TABLE exercise_metrics (
 
 -- Gamification / Achievements table
 CREATE TABLE achievements (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT REFERENCES users(id) ON DELETE CASCADE,
     title VARCHAR(255) NOT NULL,
     description TEXT,
     earned_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
@@ -39,7 +39,7 @@ CREATE TABLE achievements (
 
 -- Streaks (Gamification feature)
 CREATE TABLE user_streaks (
-    user_id INTEGER PRIMARY KEY REFERENCES users(id) ON DELETE CASCADE,
+    user_id BIGINT PRIMARY KEY REFERENCES users(id) ON DELETE CASCADE,
     current_streak INTEGER DEFAULT 0,
     longest_streak INTEGER DEFAULT 0,
     last_active_date DATE
